third_party/blink/web_tests/http/tests/js-self-profiling/external-script.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>

  <script src="resources/profile-utils.js"></script>

  <script src="resources/external-script.js"></script>
</head>
<body>
  <script>
    const BASE_URL = `${window.location.origin}/js-self-profiling`;

    promise_test(async t => {
      const trace = await ProfileUtils.profileFunction(function trampoline(sample) {
        externalScriptFunction(sample);
      });

      assert_true(ProfileUtils.containsResource(trace,
        `${BASE_URL}/resources/external-script.js`,
      ), 'external resource is included');

      const expectedTrampolineFrame = {
        name: 'trampoline',
        resourceId: trace.resources.indexOf(
          `${BASE_URL}/external-script.html`
        ),
      };
      const expectedExternalFrame = {
        name: 'externalScriptFunction',
        resourceId: trace.resources.indexOf(
          `${BASE_URL}/resources/external-script.js`,
        ),
        line: EXTERNAL_SCRIPT_FUNCTION_LINE,
        column: EXTERNAL_SCRIPT_FUNCTION_COLUMN,
      };

      assert_true(ProfileUtils.containsFrame(trace, expectedTrampolineFrame),
                  'trampoline function included');

      assert_true(ProfileUtils.containsFrame(trace, expectedExternalFrame),
                  'external script function included');

      assert_true(ProfileUtils.containsSubstack(trace, [
                    externalScriptFunction,
                    expectedTrampolineFrame,
                  ]),
                  'stack exists with external script function');

    }, 'external script function details');
  </script>
</body>
</html>
